style: fixed coding style and move class NormTC
